Add the check for errors while running the program
In this way it should be easy to run the program for a long period of time and check if in the meantime some error occurs. In case of errors, the algorithm halts but the program doesn't stop, but until now it wasn't printing anything to the standard output since it was redirected to the log file.